memory-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-mcp"
version = "0.1.0"
description = "MCP server for AI long-term memory - Let AI remember across sessions!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "sentence-transformers>=2.0.0",
    "rank-bm25>=0.2.2",
    "sudachipy>=0.6.10",
    "sudachidict-core>=20260116",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "jurigged>=0.6.0",
]

[project.scripts]
memory-mcp = "memory_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_mcp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
per-file-ignores = {"src/memory_mcp/server.py" = ["E501"]}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
